rename: change all references to gridgopher
